Re-structured PP Strip, 16-bit component bugs, red/green/blue global brightness, type-mismatch warnings
• Re-structured PPStrip to convert RBG values of various types to a table of bytes as these values are presented, rather than when [PPStrip serialize] is called.  After this, the code that presented the values (in uint8s) ran twice as fast, and [serialize] probably ran 100x as fast since all it does now is copy a block of bytes.

• Fixed recent code written to handle 16-bit RGB components (not tested yet since we don't have the 16-bit hardware yet!)

• Expanded the global brightness properties in various classes to handle separated brightness adjustments for red, green, and blue.

• Fixed a few type-mismatch warnings, particularly ones where an unsigned int was compared to a signed int.

+++ PixelPusher-objc/Source/PPCard.m
@@ -175,7 +175,7 @@ INIT_LOG_LEVEL_INFO
 	float powerScale = PPDeviceRegistry.sharedRegistry.powerScale;
 
 	const NSUInteger nStrips = self.pusher.strips.count;
-	int32_t stripIdx = 0;
+	NSUInteger stripIdx = 0;
 	const int32_t requestedStripsPerPacket = self.pusher.maxStripsPerPacket;
 	const int32_t stripPerPacket = MIN(requestedStripsPerPacket, (int32_t)nStrips);
